Build manifest for org.mypaint.MyPaint
This will build current stable MyPaint v1.2.1 .

MyPaint still needs to be ported to use the File Chooser portal, so I'm
opening access to the host files until that's fixed.

Wacom tablets won't work under Wayland.  This is fixed in MyPaint
master, so I'll be watching for next stable release to update the
manifest.  By the way, under Wayland the application starts fine, but if
you want to use a wacom you better switch to X11.  For the momment I'll
be providing a (Nightly) Mypaint outside Flathub.