Add LSP LV2 plugins extension
The lsp-plugins collection of LADSPA/LV2/VST compatible plugins use a
rare funding model. All plugins are proprietary until a donation
milestone is reached. This means some plugins can only be provided
from binary sources.

Fortunately this is not an issue, because the license does allow
unmodified re-distribution of the upstream provided binaries.

For now, this solves license compatibility uncertainties with
com.github.wwmm.pulseeffects. It also allows users to choose
whether they want to install this or not.