Try to fix build failure on ARM
On Arm, Qt MOC generates moc class for KCMAkregatorPluginsConfigFactory
inside the Akregator namespace, which is wrong, because the class is
declared in the global namespace. I assume MOC gets confused by the
`using namespace` statement and maybe by the dependence of one of the
arguments on the Akregator namespace.

To workaround the issue we move the factory declaration to the Akregator
namespace where MOC expects it.