Temporarily use self-built repomaker package because of keystore issue
Running Repomaker with Flatpak, one was not able to make a repo because
the signing of the repo's index failed.

This was caused by the fact that the default keystore type used in
flatpak is _PKCS12_. This type doesn't allow different passwords for
store and key, which turned into this error message when creating a
keystore:
```bash
Warning: Different store and key passwords not supported for PKCS12 KeyStores. Ignoring user-specified -keypass value.
```

Essentially, the key password was just dropped and instead the password
for the store was used. Later, _jarsigner_ failed when trying to unlock
the key and its store with different passwords:
```bash
jarsigner: key associated with Key Alias not a private key
```

Repomaker's behavior when creating a repo was now changed in
https://gitlab.com/fdroid/repomaker/merge_requests/204.
It sticks to fdroidserver's behavior of using the same password for both
the store and the key. With this, Repomaker on Flatpak no more has
issues when signing the index.

Because it's really bleeding edge, there isn't a package with
this fix available on PyPi so far. In the meantime, the build script will use my
self-built version of repomaker.

Fixes https://gitlab.com/fdroid/fdroid-repomaker-flatpak/issues/13.

Thanks a lot to @Persival, @eighthave and @grote who joined in and
helped me a lot on sorting this out.