Upgrade to version 1.3.5
This upgrade also contains a temporary disabling of the electron
sandbox. Since we don't ship SUID binaries in flatpaks and neasted user
namespaces are still not a thing, this is the only way to have electron
5 applications running within flatpak.

As soon as the flatpak upstream found a better solution, we'll switch to
it and remove the `--no-sandbox` option again.

Reference: https://github.com/flatpak/flatpak/issues/3044