Enable audio support
We currently do this using the `padsp` helper provided by PulseAudio
to connect Pure Data's OSS audio backend to the desktop's PulseAudio
server.

In future we should use the Pure Data JACK backend and use Pipewire
to integrate this with the desktop. This will improve latency and
should tie in with some work by Red Hat to improve support for pro
audio tools in Flatpak[1].

1. see the "Pipewire" heading at
https://blogs.gnome.org/uraeus/2019/09/23/fedora-workstation-31-whats-new/