Add Visual Studio Code Insiders build for beta stream
Based on the existing com.visualstudio.code build, containing the
release stream of VS Code, add a separate app-ID for the Insiders
stream, which is their beta release channel.

Disclaimers have been added to the app description and warning file
opened at first startup, however this should still only be distributed
in a beta channel, and not generally.