Use SDL 2
Fuse can use SDL for the UI, the audio driver and the joystick driver.

The UI code still needs to be ported to SDL 2, but the audio and
joystick drivers work just fine as they are now.

Since this Flatpak build uses the GTK UI we can switch to SDL 2 and
stop building and shipping SDL 1.