Add a manifest for jammr 1.2.8
jammr allows musicians to jam together online (https://jammr.net/).
This app contains the GPL-2.0-or-later licensed client which is a Qt
application developed by the upstream Wahjam project
(https://wahjam.org/).  I am the owner of jammr and the maintainer of
Wahjam.

The online jamming platform that the client connects to is free to use.
A Premium subscription is available offering additional features.  I
think this app fits within the scope of Flathub since it is open source
and can be used for free.

Add a manifest for jammr 1.2.8 using org.kde.Platform 5.14.  Luckily
most dependencies are already present in org.kde.Sdk.  Add PortAudio and
PortMidi modules similar to how the org.mixxx.Mixxx manifest does it.

The client needs access to ALSA, PulseAudio, and/or JACK audio devices.
It also uses networking and is a GUI application.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>