Initial working flatpak manifest.
Some parts of the manifest inspired by org.godotengine.godot.BaseApp
manifest, except that the BaseApp is managing Godot games the wrong way.
I don't want to provide a PCK as a source since it is pre-packaged data,
hence is bad for build transparency.
Instead, this manifest builds the game from the actual game source,
making it hence a much more trustable build. Also the finale binary does
not need Godot anymore.

Also adding a basic desktop file in here, but it will likely end up
upstream eventually as I want to help the developer a bit to package his
application (right now it's quite bare).

This flatpak is working fine (tested only on x86_64 so far, I'll tweak
the manifest on flathub to see how it will go on ARM) as we can confirm
after spending a good evening playing this fun game! :-)