Add a flathub.json and skip arm for now.
Godot fails to build on ARM. I'm not sure why yet, though a quick web
search indicates it might be a SSE flag to add.
Anyway for this initial flathub build, I will be happy with the x86_64
and aarch64 only builds. Let's see later to fix and reenable ARM.