Add parsec flatpak build
This commit adds the current progress on the parsec flatpak for app
submission. It should provide all needed data for a successful
submission but reviews are very welcome and improvements will definitely
be possible.

Main concern about submission: versioning. This has to be figured out
with Parsec Cloud Inc. involved as from there side versioning
information is still missing.