Rebrand Riot to Element
Riot was renamed to Element[1] by the upstream project and recieved an
update to version 1.7.0 as part of the rebranding. With some discussions
in the project there is the expectation to **not** change the
application ID of the flatpak just because the application was
renamed.[2] Therefore this patch only updates the visible indicators
but keeps the app-ID identical.

This patch renames all visible indicators from Riot to Element, upgrades
Element to version 1.7.0 and fixes the build system were needed.
Otherwise the flatpak itself stays the same, we still repackage the
upstream `.deb` as well as using the `im.riot.Riot` app-ID.

[1]: https://element.io/blog/welcome-to-element/
[2]: https://github.com/flathub/im.riot.Riot/issues/114