Enable file-picker support in zypak
This patch will force electron to use the gtk portal as filepicker. This
allows us to drop all directory access for the app without losing access
to file. This will fix drag-drop support as well as saving and uploading
files in Riot without requiring any special permissions.

Thanks a lot @refi64 for implementing this feature into the zypak!