Add World Possible ScriptLauncher manifest
This is a copy of the upstream manifest
(https://github.com/endlessm/worldpossible-scriptlauncher/) with the
branch changed from main to the 1.0 tag so the build is reproducible (or
at least that part of it is). This copy of the manifest also has the
branch removed since that defaults to "stable" on Flathub, and removes
the "V=1" build option since that's set in the runtime.

ScriptLauncher is an app we at Endless OS Foundation (endlessos.org)
created for World Possible (worldpossible.org) to assist them in their
administration of Endless OS computers that they have deployed in
various places (mostly for educational purposes in offline
environments). The function of it is to execute a Bash script as root
without requiring use of the Command Line Interface (though it can also
execute other things like Python scripts provided they have correct
permissions set and a correct shebang line). The reason World Possible
needs this is that they often have to make changes to deployed computers
such as installing or removing content, and the people available to push
the changes to each computer are not always experienced with the Linux
CLI.

Some consideration has been given to deploying changes with a technology
such as Ansible, but for now they are using shell scripts.

This utility requires an administrator password be entered to execute
the script as root, so it does not provide any privilege escalation.

It also has a few other features such as showing the output from the
script, but it is basically at MVP status.