Change workspace directory.
By default, configuration files are stored in
`~/PortfolioPerformance/workspace`, but the Flatpak specification
recommends that ["retaining and sharing configuration with non-Flatpak
installations is to be
avoided"](https://docs.flatpak.org/en/latest/sandbox-permissions.html).
This patch changes this to the `workspace/` directory inside
[`$XDG_CONFIG_HOME`](https://docs.flatpak.org/en/latest/conventions.html#xdg-base-directories),
which is `~/.var/app/${FLATPAK_ID}/config/` in Flatpak.

Note: Once buchen/portfolio#1922 is fixed, this will no longer be necessary.