Initial flatpak packaging
This packaging uses the currently latest tarball as a base, and adds
everything needed for the application to work as a flatpak, along with a
new hi-res icon.

The changes and additions made here would be better placed upstream in
the long term, but for now they make the flatpak work.