Provide Voikko as a bundled extension for LibreOffice
...to fix <https://github.com/flathub/org.libreoffice.LibreOffice/issues/122>
"Add Voikko for Finnish spell check".  (There appears to be no good reason that
the official Finnish Spellchecker extension for LibreOffice, see
<https://extensions.libreoffice.org/en/extensions/show/finnish-spell-checker-and-hyphenator-voikko>
and <https://www.puimula.org/htp/ooo/voikko-win/5.0.2.20170827/voikko.oxt>, does
not provide the necessary files to work on Linux in general, and thus in the
LibreOffice flatpak in particular, see the mail thread starting at
<https://lists.puimula.org/pipermail/libvoikko/2021-August/000955.html>
"[libvoikko] Choice of native code inclusion in voikko.oxt for LibreOffice".
But as long as that general issue is not addressed, provide it as a
bundled flatpak extension, similar to how Linux distros provide it as a
bundled extension in whatever distro package formats.)

This is based on (and a partial copy of)
<https://github.com/jpakkane/org.libreoffice.LibreOffice/commit/b98d202f4deb6f7f13a27bf0b770dfe73b1de2f5>
"Add Finnish spell checker and hyphenation. Closes flathub#122." (cf.
<https://github.com/flathub/org.libreoffice.LibreOffice/pull/155> "Add Finnish
spell checker and hyphenation. Closes #122."), but provides its functionality as
a bundled extension rather than as a core part of the
org.libreoffice.LibreOffice app.  Also, in line with Fedora's
<https://src.fedoraproject.org/rpms/libvoikko/c/374ff8b09fd2ff289e7777ab22dcd21bce4ee1f5>
"Update to 4.3" (as used by Fedora's libreoffice-voikko equivalent of this
bundled flatpak extension), --disable-hfst to avoid a bunch of dependencies.

The upstream pythonpath/VoikkoHandlePool.py expects the native voikko dynamic
library (libvoikko.so.1 on Linux) in a voikko/ sub-directory whose name is based
on the Python platform.system() ('Linux') and platform.architecture() (a tuple
like ('64bit', 'ELF'), but which oddly lacks the CPU architecture) values.  To
simplify the build script and not have to guess the appropriate Python values
for the sub-directory name there, libreoffice-voikko-path.patch changes that
logic to a sub-directory just named "Linux".  (Which is fine as each build of
this bundled extension only contains native code for one specific architecture
anyway.)

This relies on
<https://github.com/flathub/org.libreoffice.LibreOffice/commit/2add5864d55a5cbe376da8a342549a8dbf46387b>
"Merge pull request #158 from flathub/extensions:  Add an
org.libreoffice.LibreOffice.BundledExtension extension point".