Adapt for Threema according to Threema Web Electron documentation
We don't run the `electron:package:linux:deb:consumer` script (or `linux:rpm`)
since we don't really want to generate a .deb package.  The other references to
`linux:deb` are a bit confusing, but in fact, for `electron:dist` and the patch,
there is no difference b/w `linux:deb` and `linux:rpm`.

We also manually list the sources of the git submodule at the specified tag.

Upstream doc:
https://github.com/threema-ch/threema-web-electron/blob/release-1.0.3-latest/README.md