Updated to a newer commit that uses an updated Manifest file, which adds all the necessary information.
Removed the following permissions:

- "--filesystem=host" because files are all accessed through portals. Turns out this was not necessary after all.
- "--share=network" because the ability to download avatar pictures is non-essential and it's implementation must be changed to use portals upstream, at the Kirigami framework. I have no control over it.
- "--device=all", because I have yet to implement the features that make use of devices. That should happen later this year. I'll come back and update the Manifest then.

I did not update the runtime version to "5.15-21.08" because I haven't been able to debug why it crashes on application start. There is an issue preventing me from installing KDE's SDK.Debug package from Flathub, and without it I can't proceed on this system. I'll try to set up a VM later and see if it works there.