Update to PSPP 1.6.2 to fix PSPP icons.
PSPP 1.6.1 switched various application IDs to "org.gnu.pspp", but some of
them were missed, most importantly the ones in the embedded icon resources,
which meant that the icons didn't display as intended.  This fixes that
problem and a few other related to application IDs, plus a couple of other
bug fixes.