Upload files from InfiniteShooter-Shortcut
Instead of having these files as a .zip or tarball that's pulled from the web, it's probably easier to have them as local files and just get InfiniteShooter.pck from the Internet. That way, it should be in line with some of the practices other packages are using while making it easier to make modifications to the flatpak and reducing dependence on external sources.