added io.github.beamerpresenter.beamerpresenter manifest
The source points to a commit shortly after version 0.2.3 of
beamerpresenter. The difference to v0.2.3 lies only in small
adaptions required for flatpak.

This manifest defines building for Qt 5.15. It is also possible to
build for Qt 6.3 with minor changes (adapt version numbers, enable
Qt 6 in poppler and disable Qt 5 in poppler).

This manifest defines one possible configuration of BeamerPresenter.
Other configurations with other PDF engines have not been tested with
flatpak but might become available in the future.