Revert to --socket=x11
This is kind of special for FreeRDP:
By default we use a shell script wrapper which decides the binary
to use depending on environment. (wlfreerdp or xfreerdp)
This might be overruled by adding --command=xfreerdp to force the
use of xfreerdp on a wayland session.
Now, with --socket=fallback-x11 xfreerdp will not find a working
X-server and fail to start, whereas --socket=x11 works like a charm.