supersedes org.racket_lang.DrRacket
Based on the feedback from the Racket community, the entry point is preferred to
be `racket`, which is consistent with other Racket distributions. However, using
the new entry point with the old application id means that `flatpak run
org.racket_lang.DrRacket` opens a repl, which is very likely to throw users off.

Therefore, the new application is created. The old application will later be
declared EOL.