remove golang build failure workaround
This was to work around permission problems with the golang
environment. At first i just couldn't build the flatpak at all. it
would fail near the very end with:

Error: module supersonic: unlinkat: Permission denied

I found this hack on:

https://www.reddit.com/r/flatpak/comments/n8y3he/comment/gxla9og/?context=3

One of the build log failures looked like:

https://paste.anarc.at/publish/2023-04-13-WKiMhZ8aicJYxQkrjVQTgaioagwkf5EjSRY42P0TkZw/clipboard.txt

and back then the YAML file looked like:

https://paste.anarc.at/publish/2023-04-13-WKiMhZ8aicJYxQkrjVQTgaioagwkf5EjSRY42P0TkZw/com.github.dweymouth.supersonic.yml

I've had to squash so many of those things to even get the app to compile that I haven't dared to remove this yet, but maybe it's worth a try!