Add icon
The source tarball contains a 64×64 icon but Flathub also requires
128×128.

To make the 128×128 icon, I started from that icon before
https://github.com/fragglet/sdl-sopwith/commit/049177143cf253d1974ae76976d79b7b5b4c1a3b
(which added a drop shadow), scaled it up to 128×128, and then reapplied
a drop shadow in the GNU Image Manipulation Project that looks somewhat
visually similar to the 64×64 one. It wasn't an exact science.