use new tagged release that doesn't require changing default build-options
The problem that required disabling exceptions was a place where std::exp() was called on a NaN value - the code checked if the result was NaN, which when assertions were enabled meant that exp() would throw before I could deal with it. Now the argument is checked before that happens, and the assertions don't cause problems in my test suite.