Switch from wildcards to specific filename in build commands
This commit changes the means of matching the file for the 'mv' command in the 'build-commands' section of the 'online.hatsune_miku.tachidesk-vaadinui.json' from a wildcard pattern to a specific filename. This change also adds the 'dest-filename' field in the JSON to specify the file name from the download URL. This makes the command more specific.