disabled wayland support as it introduces bugs with the input
alt+scroll let you change how long is the note placement, but with the wayland native window, it breaks. Yes, you can change that in the left bottom corner, but it's much more reliable to use that shortcut. I contacted the devs and community and eventually we will try to fix it, because the wayland window was pretty nice looking.