refactor: avoid checking the exit code
Checking for the exit code is not perfect, it might be a SIGABRT, you should just let it fail.

If someone wanted a native Wayland window they should get it or see the failure instead of a fallback. If it fails they can make a conscious choice of switching to Xwayland.

https://github.com/flathub/flathub/pull/5342#discussion_r1658921987