Change the source location from GIT to TGZ
If the xca buildsystem detects a git repository it extends
the version number in the "About dialog" by the commit hash
and informations about local changes (patched metainfo.xml).

This is ugly and suggests an untagged build, which is actually not true.

Switching to the tar.gz repository resolves this.