Add Pidgin3
This is still technically unreleased, but flatpak is one of our preferred
installation methods so we are using a development tarball.

We debated about using flathub beta, but flathub proper is probably the way to
go, but we are going to have many releases before we have a stable release.