build from source
Change to build Chunker from source instead of
unpacking release bundles.

This also includes a patch to make electron-builder
produce only an unpacked output, since keeping DEB
and AppImage targets would force electron-builder
to download FPM, which is unable to run in Flatpaks
due to a hard requirement on libcrypt.so.1.